server/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
firebase-admin
cloudinary
django-cloudinary-storage
Pillow
requests
PyJWT
mongoengine
pymongo
dnspython
blinker